prototype_website_chatbot/requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
streamlit
langsmith
grandalf
langchain-groq
langchain-text-splitters
lark
langchain-huggingface
langchain_ollama
langchain_chroma